https://www.hackerrank.com/challenges/matching-questions-answers
Implemented a similarity function to identify the matching
sentences(from paragraph) for each answer. The same function is used
again to match the questions with the corresponding sentences. The
similarity was measured using a function that takes the count of words
and Inverse Document Frequency as factors. Stop words are removed. The
code passed in 13 out of 14 test cases.